Rename projectionComponent.js to projection.js to match convention
